pyproject.toml
[project]
name = "gangboss-discordbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.12.0",
    "apscheduler>=3.11.0",
    "discord-py[voice]>=2.5.2",
    "python-decouple>=3.8",
    "pytz>=2025.2",
]
